Allow the usage of a string or int as a value in portMappings
This allows the use of specifying either (for example) "80" or 80 within
portMappings of the Kedge file.

Due to it either being an array of int64 or an array of strings, this
code adds the functionality of using either/or by overriding the default
unmarshaler (UnmarshalJSON) for a custom one.

The example has also changed (for the integration test) so that 3306 can
be specified instead of "3306" for mariadb.